hubertsia_project3/requirements.txt
Flask
Flask-SocketIO
pymongo
eventlet
python-dotenv
requests